backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
azure-storage-blob==12.19.0